add stdint

+++ workshop/hardware/expressions-and-control-flow/workshop/CodeBlocks/15.c
@@ -1,7 +1,8 @@
 #include <stdio.h>
+#include <stdint.h>
 
 int main() {
-	int u = 123;
+	uint8_t u = 123;
 	// print 'Hooray!' if the number is bigger than 100
 	return 0;
 }
